CRUD Methods get user id from httpcontextacessor

+++ Services/CharacterService/CharacterService.cs
@@ -27,9 +27,10 @@ namespace dotnet_rpg.Services.CharacterService
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
-            _context.Characters.Add(character);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUseriId());            _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             serviceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUseriId())
                 .Select(c => _mapper.Map<GetCharacterDto>(c))
                 .ToListAsync();
             return serviceResponse;
@@ -41,11 +42,20 @@ namespace dotnet_rpg.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
+                Character character = await _context.Characters
+                .FirstOrDefaultAsync(c =>  c.Id == id && c.User.Id == GetUseriId());
+                if(character != null)
+                {
                 _context.Characters.Remove(character);
                 await _context.SaveChangesAsync();
-
-                response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                response.Data = _context.Characters
+                    .Where(c => c.User.Id == GetUseriId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }else{
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
+                
             }
             catch (Exception ex)
             {
@@ -69,7 +79,8 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUseriId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -81,8 +92,11 @@ namespace dotnet_rpg.Services.CharacterService
             try
             {
                 var character = await _context.Characters
+                    .Include(c => c.User)
                     .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
 
+                if(character.User.Id == GetUseriId())
+                {
                 character.Name = updatedCharacter.Name;
                 character.HitPoints = updatedCharacter.HitPoints;
                 character.Strength = updatedCharacter.Strength;
@@ -93,6 +107,13 @@ namespace dotnet_rpg.Services.CharacterService
                 await _context.SaveChangesAsync();
 
                 response.Data = _mapper.Map<GetCharacterDto>(character);
+                }else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
+
+               
             }
             catch (Exception ex)
             {
